requirements.txt
alembic==1.3.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
httpie==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
Pygments==2.5.2
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.11
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
psycopg2==2.7.3.1
gunicorn==19.7.1